Set TRACEBACK=yes by default if it isn't explicitly passed
We have modified the Makefile_header.mk file to set the GEOS-Chem
Makefile option TRACEBACK to "yes" if it is not already passed as a
command line argument.  This will ensure that GEOS-Chem will print
out the error stack trace (i.e. the list of called subroutines and
line numbers at the point of an abnormal exit).

Seb Eastham (Harvard) suggested this fix.  Having the error stack
printout will make it easier for GEOS-Chem users to diagnose and
report issues.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>