Tell user to look in HEMCO log file for error messages
We have modified routine ERROR_STOP in GeosUtil/error_mod.F to
display an optional string with further instructions for the user.

In the routines in hcoi_gc_main_mod.F90, we use this extra
argument to add "HEMCO ERROR: Please check the HEMCO log file for
error messages!" to the standard GEOS-Chem error stop message.
This should help GC users find errors that may not be printed
to the stdout stream (or log file).

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>