Avoid array out-of-bounds error when reading jv_spec_mie.dat with brown carbon option
When using the brown carbon option, an array out-of-bounds error would
occur for TITLEAAt because that array only has 7 dimensions, but was
indexed with dimensions 57-63. The temporary arrays used in RD_MIE for the
brown carbon entries weren't necessary, so they have been removed. Instead,
if the brown carbon option is on, those entries are read and directly
overwrite the entries for organic carbon.

The write statements the end of RD_MIE have also been updated so that the
entire title lines from jv_spec_mie.dat are printed to the log file.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>