Skip lat-lon grid computation when using an external grid
Bug fix submitted by Seb Eastham: "Internal computation of the grid gives
invalid results when running in a grid-independent mode. This call is
therefore skipped during initialization when an external grid is defined.
The grid parameteres XMID, YMID, XMID_R, YMID_R and AREA_M2 in grid_mod
must be set externally when this is the case."

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>