Now apply P/L of NITs throughout the troposphere
Back when dry deposition was applied in the CHEM_NIT routine, we had an
IF block that restricted applying the P/L of NITs to within the
boundary layer.  This is now deemed to be unnecessary, so we have
removed it.

We have also rewritten the CHEM_NIT routine to remove redundant variables,
also for the dust uptake tracers.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>