Remove drydep use of State_Met%RADSWG for GCHP when using GEOS-FP/MERRA2
This GC_BLEEDING_EDGE update is concurrent with a GCHP repository
update that sets State_Met%SWGDN rather than State_Met%RADSWG if
using GEOS-FP/MERRA2. That change is for consistency with the
GEOS-Chem classic convention of storing radiation in State_Met%SWGDN and
not State_Met%RADSWG when using those met sources.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>