Force transport timestep to be considered in smallest timestep calc
The transport timestep is now always considered when calculating
the smallest timestep. Previously, turning off transport would
result in the dynamic step being extended to the full chemistry
timestep, which would in turn change the time at which chemistry
was performed. This fix ensures that chemistry will always be
performed at the same time, whether or not transport is enabled,
at the expense of a small increase in overhead when transport is
disabled.

Signed-off-by: Sebastian Eastham <seastham@fas.harvard.edu>