Check for advected tracers only in input.geos
There are multiple checks in input_mod designed to ensure that, if
a certain species is present, the code designed to work with it is
also present. This is a problem for GCHP, which always defines all
species based on the KPP mechanism, even when the simulation doesn't
need them. To get around this, the checks in input_mod now only look
for advected tracers (i.e. those in input.geos), which is consistent
with the v10 behavior of GC.

Signed-off-by: Sebastian Eastham <seastham@fas.harvard.edu>