Fixed typo in #ifdef in routine INIT_WINDOW
In the nested-grid routine INIT_WINDOW (GeosCore/tpcore_window_mod.F90),
we changed the #if defined( GEOS5 ) to #if defined( GEOS_5 ).  This will
now correctly define the DLON parameter (which had been set to zero
inadvertently).

We compile GEOS-Chem with the C-preprocessor command -DGEOS_5, which
sets the GEOS_5 switch (but not GEOS5, which was incorrect).

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>