Add initial code modifications for PAN updates from Emily Fischer
Emily Fischer provided us with updates to the PAN simulation. Her code
updates were based on v9-01-01, but had been merged with v9-02h. To avoid
duplicate code and other errors, we have manually brought the PAN updates
into v11-02a. A good portion of the updates were made in the emissions
routines, so these updates will need to be implemented via HEMCO.

For more information on the updates to the PAN simulation, see:

  http://wiki.seas.harvard.edu/geos-chem/index.php/PAN

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>