Now replace COMPILER with COMPILER_FAMILY in the Makefile_header.mk
Further refine the build sequence to replace the COMPILER makefile
variable with COMPILER_FAMILY.  COMPILER_FAMILY is now set depending on
the value of the FC environment variable.

  When FC = gfortran,              COMPILER_FAMILY = "GNU"
  When FC = ifort|mpif90|mpifort,  COMPILER_FAMILY = "Intel"
  When FC = pgfortran|pgi|pgf90,   COMPILER_FAMILY = "PGI"

The COMPILER_FAMILY variable is used in IF statements to pick the
proper compilation and optimization options.  This allows for the
compile command to be e.g. "mpif90: but the compiler family to be
"Intel", in order to select the proper behavior.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>