Bug fix: include partial hour in current time used for State_Met%suncos
Prior to this commit, the current hour rather than the current time was
used to calculate cosine of current time solar zenith angle. This resulted
in a constant State_Met%suncos for all timesteps within the same hour.
State_Met%suncosmid (cos(SZA) at the mid-point of each timestep) was not
affected.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>