Restore P(OH) in ND22 diagnostic for tropchem simulations
Prasad Kasibhatla pointed out that production of OH from O3 photolysis
(index 6) in the ND22 diagnostic is saved out as all zeroes in GEOS-Chem
v11-01. This diagnostic quantity is only saved out for the tropchem and SOA
mechanisms (i.e. simulations that do not use UCX). In the Standard and UCX
mechanisms, J(O3P) is saved out for index 6 of ND22. The P(OH) quantity was
deactivated during the implementation of FlexChem and was not restored. To
restore P(OH), we add ZPJ for RXN_O3_2a and RXN_O3_2b to AD22 in routine
PHOTRATE_ADJ, following what was done in the old calcrate.F.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>