Remove PASV from species database and use passive tracer module instead
The PASV species was added early in v11-01 to include a passive tracer in the
Rn-Pb-Be benchmarks. Since then Christoph has added a passive tracer module
that easily allows for the addition of multiple passive tracers. To avoid
redundant code, the PASV species has been removed here and will instead be
declared in the Passive Tracer Menu of input.geos.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>