Add fixes for UCX CPP switch in Makefile_header.mk
By default, UCX should be set to y when users compile with CHEM=Standard.
In Makefile_header.mk, we now explicitly check for CHEM=Standard and set
UCX accordingly. By default, if CHEM=... is not passed, KPP_Chem=Standard
(and UCx=y) will still be set but this should be a temporary workaround
until we can add an option for the specialty simulations that do not
require KPP.

The comments in help/Makefile has also been updated here to clarify the
default settings for the various chemical mechanisms.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>