Add fixes for 3D biomass burning emission diagnostics
Biomass burning emissions from GFED and FINN are now 3D to allow for
65% of biomass burning emissions to be emitted into the boundary layer
and 35% to be emitted into the free troposphere. The biomass burning
diagnostics in ND07, ND28, and ND29 have been updated here to pass
LLPAR for number of levels when retrieving these emissions from HEMCO
and writing to the bpch diagnostics.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>