Now use the FZ array to compute the ND26 vertical flux diagnostic
Ilya Stanevic (U. Toronto) says that the FZ array should be used to
compute the ND26 vertical mass flux diagnostic, similar to how FX
is used for ND24 and FY for ND25.  Initial evaluation shows a near-
identical match to the accumulated mass for Rn, Pb, Be species, but
not for PASV.  We are still investigating.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>