Update CH4 and Hg emission arrays to allow for 3D biomass burning emissions
In the PAN updates, the GFED and FINN biomass burning extensions were
updated to distribute 65% of emissions into the boundary layer and 35%
into the free troposphere. This update also applied to species CH4 and Hg0.
As a result, the biomass burning emission fields returned by HEMCO for
BIOMASS_CH4 and BIOMASS_HG0 are 3D so the emission arrays (CH4_EMIS and
HG_EMIS) that they are copied into must be updated include a dimension for
level. The diagnostic arrays AD58 (CH4 emissions) and AD03/AD03_nat (Hg
emissions) were also updated to include a dimension for level.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>