Only define OCPI/OCPI anthropogenic emissions diagnostics when species are defined
The SOA-SVPOA simulation was crashing in DIAGN_ANTHRO when trying to
create the anthropogenic emissions diagnostic for OCPI and OCPO. In
that simulations OCPI and OCPO are replaced by POA1 and POG1. To avoid
the error, we now make sure id_OCPI and id_OCPO are set before creating
these diagnostics.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>