Update CO2 emissions unit conversion for consistency with advection pressure
A pressure update was introduced in v11-01h that corrected an advection
issue in which tracer mass was transported relative to moist rather
than dry air. To conserve mass throughout the model, all unit conversions
were updated to use a dry pressure difference consistent with the new
values used in advection. This replaced the previous method which
calculated dry pressure difference per level directly from moist edge
pressures and GMAO specific humidity.

Meemong Lee discovered that a CO2-only unit conversion for carbon
dioxide emissions continued to use the old method for converting units.
This update changes the unit conversion to be consistent with the
mass-conserving unit conversions that are consistent with advection.

Note that both methods are scientifically valid but only one enables
mass conservation when advection is enabled.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>