Fix error checks for LSOA to indicate it is for the complex SOA scheme
LSOA is now used to turn on/off the complex SOA scheme. The error checks
have been updated to specify this. A check has also been added to make
sure the simple SOA species (SOAP and SOAS) are not defined when the
complex SOA scheme is used to avoid double counting.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>