Change carbon_mod to prevent out-of-bounds error
A debug message in carbon_mod was hard-coded to check values
at location (20,33,2) and (20,33,10). However, this would
cause a problem in MPI environments, where the domain of
an individual core may be smaller than this. A check has
been put in place to ensure that the debug location is
within bounds.

Signed-off-by: Sebastian Eastham <seastham@fas.harvard.edu>