Restore ND38 diagnostic - this code requires further investigation
The parallel loops in DO_CONVECTION that use DIAG38 are problematic
in that they cause simulations to crash when debug flags are turned on
(or Gfortran, which has stricter flags, is used). The problematic code
is uncommented here and we will look into a fix for this issue later
in v11-02c.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>