Comment out problematic write statements in tomas_mod.F
Write statements with format codes 116 and 117 in tomas_mod.F were
causing issues in TOMAS simulations because 116 and 117 are never
defined. Format statements for these values could not be found in the
TOMAS code dating back to the original implementation in v8, so they
have been commented out for now to avoid issues. The print statement
immediately preceding these write statements seems to do the same thing
anyway.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>