Now only test for advected species > 0 in routine SET_MINIT
In routine SET_MINIT (in GeosCore/strat_chem_mod.F90) there was an error
check if the sum of State_Chm%Species was negative.  But because the
non-advected species are currently negative before chemistry (i.e. have
a MW of -1), then this was producing spurious error messages.  An error
message was being printed but the run was not being stopped.

We have therefore made the following changes:

(1) Only check if the sum of advected species (not all species) is
    nonzero.  If so, then print an error message and exit SET_MINIT.

(2) Put an error trap in routine INIT_STRAT_CHEM that will exit the
    routine if SET_MINIT returns failure.

(3) Put an error trap in main.F which wills top the run if the error
    from SET_MINIT propagates all the way up.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>