Bug fix: skip DustAlk IF block unless this extension is activated
In HEMCO/Extensions/hcox_dustginoux_mod.F, there was an error. The
HcoIDsAlk array will be unallocated when the DustAlk extension is turned
off.  Trying to access HcoIDsAlk when it is unallocated will of course
cause a seg fault error.

We added an error trap to shunt executon of this block of code unless
the DustAlk extension is activated in the HEMCO_Config.rc file.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>