Avoid seg fault when ND24, ND25, ND26 are off for nested grid sims
When ND24, ND25, or ND26 are turned off, the MASSFLEW, MASSFLNS, and
MASSFLUP arrays are sized (1,1,1,1).

When we point to these arrays with window indices I0_W1, J0_W1, etc.
(which are all larger than 1), this will cause a segfault.

The fix is to just point to the whole array (with no indices) if
ND24, ND25, and ND26 are turned off.  A kludge to be sure, but the
bpch diagnostics are heading out the window very shortly.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>