Add History modules for new netCDF diagnostics; also fix ProTex output
History/*_mod.F90: These are the new History modules, which keep the
linked lists for the new netCDF diagnostics.  They are compiled but
at this point have not been connected to the rest of the code yet.

Makefile_header.mk and GeosCore/Makefile: Add code to compile the
modules in the History/ folder.

In the doc/ folder, we have added Makefile_HistoryDoc.mk, gc_history.P,
and gc_history_intro.P so that we can send the header comments from
the History/*.F90 modules to a PDF file via ProTeX.

All other files in this commit have had typos fixed in the ProTeX
subroutine headers that were preventing the documentation from being
built.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>