Now initialize list of History collections (if NC_DIAG=y)
When compiling with NC_DIAG=y, GEOS-Chem will read the HISTORY.rc file
and use that information to create a linked list structure for each
diagnostic collection (and the fields within each collection).

Right now the list of collections is populated but nothing further is
done.  More work to follow.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>