Error checks for STATE_PSC and HEMCO_RESTART in ESMF environments
When running in an ESMF environment, we expect to get STATE_PSC
and HEMCO_RESTART variables from the internal state and not from
a HEMCO restart file. For safety's sake, GC now throws an error
if either of these variables are set to true with ESMF. We also
now always default to STATE_PSC = false if it is not specified
and we are in an ESMF environment.

Signed-off-by: Sebastian Eastham <seastham@fas.harvard.edu>