Connect data pointers from the registry to fields of each collection
Made several modifications so that each HISTORY ITEM belonging to a
data collection can attach the pointer to its data source, by querying
the appropriate state registry.

In other words, for diagnostic "CHEM_SPC_NO", we attempt to look up the
registered field "SPC_NO" in the State_Chm%Registry object.  We get
the various metadata, as well as the pointer to the data slice
State_Chm%Species(:,:,:,id_NO).

Similarly, for diagnostic MET_U10M, we attempt to look up the registered
field "U10M" in the State_Met%Registry.  If it is found, we again
return the metadata and a pointer to the data, etc.

Modified REGISTRY_LOOKUP to throw an error if the requested variable
cannot be found within a registry.  Also created registry_params_mod.F90
in the Headers/ folder to contain kind parameters for querying the pointer
data types.  REGISTRY_LOOKUP can now return the dimensions of the
source data array.

Modified routine HISTITEM_CREATE in order to pass a pointer to the data
source via the argument list.  This allows us to look up a diagnostic
name in a registry, get a pointer, and pass it to the HISTORY ITEM.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>