Now create a netCDF file for each collection using ncdf_mod.F90
We have added more fields to the HistContainer derived type:

 nX, nY, nZ, IsFileDefined, xId, yId, zId, tId

which will facilitate creating netCDF output with the History
component.  nX, nY, and nZ are needed to create the dimensions of the
netCDF file (the time dimension is UNLIMITED), and the *Id variables
return the netCDF Id's for each dimension.  The IsFileDefined sets a
flag if we have already called netCDF define mode.

Also: in main.F, we suppress calling History_Write if it is the first
timestep of a simulation (we can revisit this later).

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>