Add several new fields to HistItem and HistContainer types
We have now added the DimNames field to the HistItem type (in module
histitem_mod.F90).  DimNames may have the values: xyz, xy, yz, xz, x,
y, or z.  This will allow us to better choose which dimensions to use
when allocating the data array.  (NOTE: The time dimension will
always be "undefined", so DimNames only refers to the spatial
dimensions.)

We have also added NcXDimId, NcYDimId, NcZDimId, NcTDimId to the
HistItem type.  These will keep track of the netCDF dimension indices
for each history item.

We have also added FileCloseYmd, FileCloseHms, ReferenceYmd,
ReferenceHms, and CurrTimeSlice to the HistContainer object.  The
ReferenceYmd and ReferenceHms are used in the unit string for the
netCDF "time" index variable.  The CurrTimeSlice will indicate which
time value to write to dis, and FileCloseYmd/FileCloseHms are
self-explanatory.

We have made the corresponding modifications in history_mod.F90,
and history_netcdf_mod.F90.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>