Now add reference date/time to the unit string for "time"
History/history_netcdf_mod.F90: Create routine History_Expand_Date to
replace time & date tokens in a string.  This was done a couple of times
so we have now centralized it.  Also rewrote Get_Var_DimIds so that the
reference time and date is added to the unit string for "time".

History/histcontainer_mod.F90: Now initialize date/time fields with "0"
instead of with "UNDEFINED_INT.  This will print out properly.

GeosUtil/gc_grid_mod.F: Copy the latitude values into a private array
"LAT", which is then added to the registry.  It turns out that registering
the value YMID(1,:,1) will cause an array temporary when writing to
netCDF.  Also change the unit string of "time" from "GMT" to "UTC".

GeosCore/main.F: Noted that the current setup will open a new file
on the last timestep of the simulation.  This should only be done for
restart or checkpoint files but not other diagnostics.  Deal with
this later.

At this point, the netCDF file for each collection appears to have
been properly defined, and all index variables have been written properly.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>