Add a reference Julian date field to HistContainer type
We now have added a ReferenceJd field to the HistContainer type,
which will assist us in computing the elapsed minutes since the
netCDF reference time.  This will be needed to pick the proper
netcdf file timestaps.

Created History_NetCdf_Close routine in History/history_netcdf_mod.F90,
so that we can call this from multiple locations if need be.

Also reorganized the routine History_Netcdf_Define to be more
efficient.