Add bug fixes for MGLY and EOH species added in v11-02a
As with the isoprene species in the previous commit, MGLY and EOH
weren't properly being identified as KPP species, and therefore their
concentrations were not getting updated by FlexChem. This is now corrected
by adding lines for KppSpcId, KppVarId, and KppFixId for those two species
in species_database_mod.F90.

In addition, there was a line in flexchem_mod.F90 for setting the
concentration of EOH in KPP to zero that is no longer needed because EOH
is now an active species. That line has now been removed.

Finally, several species IDs have been removed from carbon_mod.F because
they're not used.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>