Add routine HistContainer_ElapsedTime to histcontainer_mod.F90
We have centralized the computation of elapsed time in minutes into
routine HistContainer_ElapsedTime (in History/histcontainer_mod.F90).
This routine returns the elapsed time in minutes with respect to
either the start of the simulation, or the netCDF reference time
(which is the model date & time when the netCDF file was created).

Also added parameter ROUNDOFF_DECIMALS to History/history_util_mod.F90,
to specify the number of decimal places we will round off to when
computing elapsed minutes.  This helps to elminate numerical noise when
taking the difference of two large Astronomical Julian Date values.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>