Added functionality to check alarm times for update, write, close
Added routine HistContainer_AlarmCheck, which tests if one of the
alarms (update, write, close) is "ringing".

Also modified HistContainer_ElapsedTime to either compute the Julian
Date if yyyymmdd/hhmmss are passed, or to take the Julian Date from
the Container%CurrentJd value if not passed.  This will avoid repeating
calculations.

Added the History_SetTime routine, which updates the ElapsedMin
and CurrentJd fields of the each HISTORY CONTAINER object at the
start of a heartbeat loop from the main program.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>