Improve alarm algorithm (but it's not 100% perfect yet)
We have made several modifications to the History routines to better
define/check if the update, file write, file close alarms are ringing.
Namely:

(1) We have added CurrentYmd, CurrentHms, UpdateIvalMin,
    FileWriteIvalMin, and FileCloseIvalMin to the HistContainer type

(2) We have modified History_SetTime and HistContainer_SetTime routines
    to update the current astronomical Julian Date with the heartbeat
    time (in days), and then from that, to compute the new Ymd, hms,
    elapsed time etc. variables.

(3) For alarm intervals less than a month, we simply compute the
    interval in elapesd minutes and add that to the current alarm time
    to compute the next time something should be done.  NOTE: Need to
    figure out intervals of longer than a month.

(4) Removed several routines that are no longer useful.

NOTE: What we might have to do is to open all the files at the start of
the simulation, then write and then close.  That would avoid having
e.g. time-averaged data for 2013-07-01 21:00 go into the data file
for 2013-07-02.  Will fix that in the next commit.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>