Fixed typo that was causing netCDF write error from commit a2e0a19
History/history_netcdf_mod.F90: In routine History_Netcdf_Init, a typo
in the indexing of the "RegistryName" variable caused a netCDF write
error (wrong dimensions of HyAm index variable).  This has now been
fixed.

Also routines Registry_Print (in Headers/registry_mod.F90) and
HistItem_Print (in History/histitem_mod.F90) now print the alignment of
the data in the vertical dimension: C if the data is defined on level
centers and E if on level edges.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>