Now save P0 to the netCDF file
Modified the registry_mod.F90, grid_registry_mod.F90, histitem_mod.F90
routines to create history items for scalar and 1D data of type REAL*8,
so that we can write the P0 parameter to the netCDF file.

Also modified the histcontainer_mod.F90 and history_mod.F90 routines
so that (1) the dimensions of the History Container are set from
the History Items, and (2) to keep History Items with the same vertical
cell position (on level centers or level edges) in the same History
Container.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>