Fix several open issues with netCDF file timestamps and metadata
GeosCore/main.F: Now pass "State_Chm%Spc_Units" to History_Write, so
that we can replace the "TBD" in the unit string with the current units
of the State_Chm%Species array at the time of file write.

Headers/errcode_mod.F90: Routine GC_ERROR now flushes the error msg
to the stdout/log file stream.

History/histcontainer_mod.F90: Added "FirstInst" and "Spc_Units"
fields to the HistContainer type.

History/history_mod.F90: Routine History_Write now accepts "Spc_Units"
as an input argument.  This is used to define to Container%Spc_Units,
so that we can pass that to history_netcdf_mod.F90.

History/history_netcdf_mod.F90:
(a) If Container%FirstInst = T (i.e. if the file is instantaneous and
    it's the first time we are creating a netCDF file), then use the
    date/time at the start of the simulation (i.e. EpochJd) as the
    reference date/time for the file.  Otherwise use the current
    date/time (i.e. CurrentJd)  This will ensure that the file names
    and timestamps for all instantaneous collection files are consistent.
(b) Now write the index variable AREA to the netCDF file as a REAL*4,
    2-D array.
(c) Replace "TBD" in unit strings with the current species units when
    creating the netCDF file.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>