Add the OnLevelEdges argument to the Lookup_State_* routines
We have added the OnLevelEdges argument to routines:

  Lookup_State_Chm  (in Headers/state_chm_mod.F90 )
  Lookup_State_Met  (in Headers/state_met_mod.F90 )
  Lookup_State_Diag (in Headers/state_diag_mod.F90)

So that we can create HISTORY ITEMS with the proper vertical position
(i.e. on level centers or on level edges).  This update will make sure
that we do not mix 3-D data with different vertical positioning
in the same HISTORY CONTAINER object.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>