Adjust UpdateAlarm for time-averaged collections
We now subtract the heartbeat timestep in minutes from the initial
setting of Container%UpdateAlarm. This will make sure that both the
instantaneous and time-averaged file collections are updated consistently
w/r/t the legacy NDXX diagnostics.  This is because the time-averaged
diagnostic routines in main.F are called at the end of each "heartbeat"
timestep (i.e. after the call to SET_CURRENT_TIME has updated the
model elapsed time).

In GeosCore/main.F, for clarity, we now call History_Update and
History_SetTime before the call to SET_CURRENT_TIME in the "heartbeat"
loop.  The code was correct before (since the History component keeps
its own time variables), but this makes it more clear to the user
that the updating is done before the timestep is incremented.

Also now make sure that some print and debug info only gets written
on the root CPU in history_mod.F90 and history_netcdf_mod.F90.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>