Bug fixes for reading collection metadata from the HISTORY.rc file
In History/history_mod.F90, we have updated the algorithm in
routine GetCollectionMetaData so that collection names with the
same characters in them (e.g. "daily" and "avgdaily") will be parsed
properly.

Also added an error trap in routine History_ReadCollectionData to
exit with an error if the collection number C as returned from
GetCollectionMetaData is undefined (i.e. = -999).

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>