Recompute file write and file close intervals if longer than 1 month
In History/histcontainer_mod.F90, we have split up the routine
HistContainer_AlarmIntervalSet into 3 separate routines, each of
which computes the alarm intervals for the "Update", "File Write", and
"File Close" operations.

The "File Write" and "File Close" alarm intervals are recomputed after
each file write and file close operation, but only if the requested
interval is 1 month or longer.  This will allow us to take into account
leap year months, etc.

NOTE: The algorithms for recomputing the file write and file close
intervals may not work properly unless you use an interval of 1 month
and 1 year.  But this should be OK for most purposes.

ALSO NOTE: We assume that we will be performing the update operation
on time intervals less than 1 month.  We typically update at the
heartbeat timestep, hourly, or daily, depending on the diagnostic
in question.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>