Add Register_MetField interface in state_met_mod to abstract registration
Metdata passed to the registry for state_met fields is now retrieved from
Get_MetField_Metadata. There are also now integer parameters for vertical
location. These are not yet used in the registry but should be
incorporated in the future based on the vloc retrieved per field within
State_met_mod.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>