Minor updates to compile with gfortran
NOTE: Subroutine NF_Dev_Var_Chunking and parameter NF_CHUNKED are
functional when compiling with gfortran but not with ifort 11 given our
default netcdf libraries. The call to NF_Dev_Var_Chunking was commented
out in a previous commit to allow compilation with ifort11. This will
continue to be commented out until a fix is developed.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>