The History component now keeps elapsed time in seconds (not minutes)
We have updated the elapsed time, alarm, and interval fields of the
HistContainer type so that they keep elapsed time in seconds instead of
minutes.  This helps us to avoid roundoff error (incurred by irrational
number timesteps such as 20 min = 0.333333_ hours) in the computation of
the elapsed time.  We simply take the integer part of the elapsed
time expression in seconds and throw away the fractional part.

We have renamed "min" to "sec" in various variables and comments
in several modules of the History folder.

NOTE: The units of the netCDF "time" variable are still kept in
minutes since the reference date & time (which is the simulation date
& time when the file is created).

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>