For now, don't include file production timestamps in History output
Prior to this commit, the History and ProdDateTime attributes contain
the date and time time at which each netCDF file (generated by the
History component) was created. This is enough to cause binary file
diffrerences -- such as in unit tests or difftests -- to fail.

The simple fix is to set the History and ProdDateTime attributes to the
null string until we can find a more robust way of differencing netCDF
files.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>