Simplify the netCDF diagnostic output file names
We have changed the default file names created by the GEOS-Chem
History component from e.g "GEOSChem.History.CollectionName.*.nc4" to
"GEOSChem.CollectionName*.nc4". Here "CollectionName" is the name of
each collection specified in the HISTORY.rc file.

This was done to avoid confusion with the GCHP output files, which
have "History" in their name.

Thanks to Lizzie Lundgren for the suggestion!

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>