Move initialization of diagnostics list to main
The diagnostics list object is used in state_diag_mod to determine what
diagnostics to allocate. It will also be used in GCHP SetServices for
determining what exports to create which means that it must be
initialized prior to the GEOS-Chem state objects. Moving the GCC
initialization to main, prior to initializating the GEOS-chem state
objects, allows initialization of it within GCHP SetServices.

Also in this commit:

1. Rename Destroy_DiagList to Cleanup_DiagList

2. Remove Print_State_Diag and Lookup_State_Diag from state_diag_mod.F90.
   Replace calls to them with calls directly to the relevant subroutine
   in Registry_Mod.F90.

3. Remove extraneous definitions of the diagnostics list object derived
   types from state_diag_mod.F90. Inclusion in this module after moving
   them to diagnostics_mod.F90 was an oversight error.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>